Added shared_ptrs for the envelopes and vectors they are composed of so they can be reused in the future. Also, they make it so only one copy of the envelopes and vectors are needed to propagate the release trigger directly to the envelopes through the native-lib.
genrateSamples of the AudioGenerator now returns a bool that is true when either of it's envelopes is finished.

Put AudioPlayer in it's own class.

Fixed envelope bugs where the release was not triggered when the play button was released and the issue where the envelope just looped from the end of the sustain to the start of the attack. Now the release is triggered when the play button is released, and the sustain loops correctly.

Added a custom play button to handle the up and down touch events.