FragmentEnvelope needed to be adjusted and looks ok now.
Working on creating the JNI function that will create an amplitude envelope. So far it extracts the integers that are enums or a function and the doubles that are arguments for each function.

Next up is creating a method that switches on the function and generates the correct segment, and another the takes in those three segments and creates the envelope.