Working on getting the envelopes graphed
Fixed some null pointer issues when getting the min and max of the envelope segments.
Drawing has to be one one the surface instead of a canvas due to ndk constraints.
Added a button to open the envelope graphs.
Got the ViewHolder for the FunctionView and the adapter for the ViewHolder implemented and set up a Recycler in MainFragment with them.
The RecyclerViewis set up with a ItemDecorator to pad FunctionViews as needed, so the amplitude and frequency envelopes stay in separate rows of the RecyclerView.