Added views to the end of both rows that will be used to allow the user to append envelopes.
Deleted AddView because it was much less work to draw the add view in the function view. Might want to create a base class for both Views so they can be used interchangeably in the RecyclerViewAdapter.

Added helper function to JNI to see if a given row and column are valid indices of the envelopes.