Some refactoring and comments
Removed the envelopes from the init function.

Changed the EnvelopeDataSource and EnvelopeRepository so they save lists of envelopes instead of individual envelopes.

Removed the envelope deleting functionality since they can be removed from the list and resaved.

Added a file where all the external JNI functions will be declared on the Kotlin side.