Creating the C++ framework for the envelope sequence builder.
Created a base class for all envelopes. It offers the functionality for getting the next envelope value, the min, the max, and the size. It also allows for reading values by index and resetting the state of the envelope.

Created an EnvelopeSegment class that is an Envelope to encapsulate a single vector and iterate through it.

Created a SegmentedEnvelope class that is an Envelope encapsulating a vector of envelopes that iterates through the envelopes sequentially.

Made the vector generation methods static so the VectorGenerator class was not needed.

Split the amplitude an frequency envelope ViewHolders.