Working on getting the custom layout manager to scroll correctly. MainFragment now uses the custom layout manager instead of the StaggeredGridLayoutManager. MainLayoutManager mostly works except for some space at the end when scrolling to the left.
Created a separate view for adding new function views.

ASREnvelope now returns a boolean when nextDouble is called rather than having to call finished to see if the envelope has ended.

Added an enum for envelope type on the C++ and Kotlin sides.

Removed the function parameter classes since they are added complexity for little benefit.

Split ASR envelope from Envelope.

Created an EnvelopeSegmentFragment to allow the user to create an envelope segment.

Added MainView base class for FunctionView and AddFunctionView so the two view's functionality could be split. Gave them different ViewHolders too.

Added helper methods for the JNI to get envelope type by index, get column number by index, and get the number of amplitude envelopes.

The MainRecyclerViewAdapter now takes a callback to call when one of the MainViews are clicked. It now also gets the number of views in the first row for the custom layout manager.