Working on saving envelopes; min and max of FunctionViews are now kept track of
Added setAmplitudeEnvelope and setFrequencyEnvelope to the AudioGenerator.

Created EnvelopeArguments to allow saving of envelopes.

Created a file with the names of all the envelope files.

The EnvelopeRepository now saves amplitude and frequency envelopes rather than just one generic envelope type.

Added a EnvelopeSegmentCache to hold the EnvelopeArguments.

Created a EnvelopeSegmentsUseCase so a EnvelopeSegmentCache could be shared across fragments.

Created a EnvelopeSegmentViewModel to clean up  the EnvelopeSegmentFragment

Added updateMinMax so FunctionView can rescale. Created a MinMax class to be propagated to all the relevant FunctionViews through the MinMaxSubject and MinMaxObserver relationship.

Cleaned out the MainActivity

Created a MainApplication to house the EnvelopeRepository.

Fixed the width calculation for the MainLayoutManager

Added a Toolbar to the main activity and added a save button to the Toolbar.