Added a dialog to get an envelope name from the user when they want to save an envelope.
Added a file to contain the names of all saved envelopes.

Created an envelope list fragment to allow the user to work on multiple envelopes.

Wrapped the amplitude and frequency envelopes into one class so they can be saved in one file.

Removed the unused buttons from the title fragment.

Removed the initial min max updates from the function views.
Instead, the adapter gets the mn max for the specified position and sets it in the min max subject so the initial update happens through the subject.

Cleaned up the main layout manager of unneeded member variables.

Added lots of comments.