Fixed a bug where the first envelope loaded was the only one to be displayed.
Added a button to play and pause the current envelope.

Fixed a bug where the saved envelope would overwrite the current envelope when popping the backstack to the envelope fragment.

Now clear the envelopes when creating a new envelope to prevent the last loaded envelope from being displayed.

Fixed a bug where playing an empty envelope caused a crash.